pyproject.toml
[tool.poetry]
name = "vandr-message-brocker-template"
version = "0.1.0"
description = ""
authors = ["Viktor Andriichuk <v.andriichuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aioboto3 = "^12.3.0"
pydantic = "^2.6.2"
httpx = "^0.27.0"
pyjwt = "^2.8.0"
ruff = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
